+++ non_catalog_apps/metroflip/scenes/plugins/myki.c
@@ -169,7 +169,6 @@ static void myki_on_enter(Metroflip* app) {
 
         // Start worker
         view_dispatcher_switch_to_view(app->view_dispatcher, MetroflipViewPopup);
-        nfc_scanner_alloc(app->nfc);
         app->poller = nfc_poller_alloc(app->nfc, NfcProtocolMfDesfire);
         nfc_poller_start(app->poller, myki_poller_callback, app);