+++ non_catalog_apps/metroflip/scenes/plugins/opal.c
@@ -283,7 +283,6 @@ static void opal_on_enter(Metroflip* app) {
 
         // Start worker
         view_dispatcher_switch_to_view(app->view_dispatcher, MetroflipViewPopup);
-        nfc_scanner_alloc(app->nfc);
         app->poller = nfc_poller_alloc(app->nfc, NfcProtocolMfDesfire);
         nfc_poller_start(app->poller, opal_poller_callback, app);