+++ non_catalog_apps/metroflip/scenes/plugins/suica.c
@@ -28,9 +28,7 @@
 
 // Probably not needed after upstream include this in their suica_i.h
 
-#define TAG                  "Metroflip:Scene:Suica"
-#define JAPAN_IC_SYSTEM_CODE 0x0003
-#define OCTOPUS_SYSTEM_CODE  0x8008
+#define TAG "Metroflip:Scene:Suica"
 
 const char* suica_service_names[] = {
     "Travel History",
@@ -309,13 +307,13 @@ static void suica_parse(SuicaHistoryViewModel* my_model) {
 
 static bool suica_model_pack_data(
     SuicaHistoryViewModel* model,
-    const FelicaSystem* suica_system,
+    const FelicaData* felica_data,
     FuriString* parsed_data) {
-    uint32_t public_block_count = simple_array_get_count(suica_system->public_blocks);
+    uint32_t public_block_count = simple_array_get_count(felica_data->public_blocks);
     bool found = false;
     furi_string_printf(parsed_data, "\e#Japan Transit IC\n\n");
     for(uint16_t i = 0; i < public_block_count; i++) {
-        FelicaPublicBlock* public_block = simple_array_get(suica_system->public_blocks, i);
+        FelicaPublicBlock* public_block = simple_array_get(felica_data->public_blocks, i);
         if(public_block->service_code == SERVICE_CODE_HISTORY_IN_LE) {
             suica_add_entry(model, public_block->block.data);
             furi_string_cat_printf(parsed_data, "Log %02X: ", i);
@@ -329,27 +327,27 @@ static bool suica_model_pack_data(
     return found;
 }
 
-static bool suica_help_with_octopus(const FelicaSystem* suica_system, FuriString* parsed_data) {
+static bool suica_help_with_octopus(const FelicaData* felica_data, FuriString* parsed_data) {
     bool found = false;
-    for(uint16_t i = 0; i < simple_array_get_count(suica_system->public_blocks); i++) {
-        FelicaPublicBlock* public_block = simple_array_get(suica_system->public_blocks, i);
+    for(uint16_t i = 0; i < simple_array_get_count(felica_data->public_blocks); i++) {
+        FelicaPublicBlock* public_block = simple_array_get(felica_data->public_blocks, i);
         if(public_block->service_code == SERVICE_CODE_OCTOPUS_IN_LE) {
             uint16_t unsigned_balance = ((uint16_t)public_block->block.data[2] << 8) |
                                         (uint16_t)public_block->block.data[3]; // 0x0000..0xFFFF
 
-            int32_t older_balance_ten_cents = (int32_t)unsigned_balance - 350;
-            int32_t newer_balance_ten_cents = (int32_t)unsigned_balance - 500;
+            int32_t older_balance_cents = (int32_t)unsigned_balance - 350;
+            int32_t newer_balance_cents = (int32_t)unsigned_balance - 500;
 
-            uint16_t older_abs_ten_cents =
-                (uint16_t)(older_balance_ten_cents < 0 ? -older_balance_ten_cents : older_balance_ten_cents);
-            uint16_t newer_abs_ten_cents =
-                (uint16_t)(newer_balance_ten_cents < 0 ? -newer_balance_ten_cents : newer_balance_ten_cents);
+            uint16_t older_abs_cents =
+                (uint16_t)(older_balance_cents < 0 ? -older_balance_cents : older_balance_cents);
+            uint16_t newer_abs_cents =
+                (uint16_t)(newer_balance_cents < 0 ? -newer_balance_cents : newer_balance_cents);
 
-            uint16_t older_dollars = (uint16_t)(older_abs_ten_cents / 10);
-            uint8_t older_cents = (uint8_t)((older_abs_ten_cents % 10) * 10 );
+            uint16_t older_dollars = (uint16_t)(older_abs_cents / 100);
+            uint8_t older_cents = (uint8_t)(older_abs_cents % 100);
 
-            uint16_t newer_dollars = (uint16_t)(newer_abs_ten_cents / 10);
-            uint8_t newer_cents = (uint8_t)((newer_abs_ten_cents % 10) * 10);
+            uint16_t newer_dollars = (uint16_t)(newer_abs_cents / 100);
+            uint8_t newer_cents = (uint8_t)(newer_abs_cents % 100);
             furi_string_printf(parsed_data, "\e#Octopus Card\n");
             furi_string_cat_str(
                 parsed_data, "::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::\n");
@@ -359,7 +357,7 @@ static bool suica_help_with_octopus(const FelicaSystem* suica_system, FuriString
             furi_string_cat_printf(
                 parsed_data,
                 "Balance: %sHK$ %d.%02d\n",
-                older_balance_ten_cents < 0 ? "-" : "",
+                older_balance_cents < 0 ? "-" : "",
                 older_dollars,
                 older_cents);
 
@@ -371,7 +369,7 @@ static bool suica_help_with_octopus(const FelicaSystem* suica_system, FuriString
             furi_string_cat_printf(
                 parsed_data,
                 "Balance: %sHK$ %d.%02d\n",
-                newer_balance_ten_cents < 0 ? "-" : "",
+                newer_balance_cents < 0 ? "-" : "",
                 newer_dollars,
                 newer_cents);
 
@@ -399,17 +397,18 @@ static void suica_parse_detail_callback(GuiButtonType result, InputType type, vo
 static NfcCommand suica_poller_callback(NfcGenericEvent event, void* context) {
     furi_assert(event.protocol == NfcProtocolFelica);
     NfcCommand command = NfcCommandContinue;
+    // MetroflipPollerEventType stage = MetroflipPollerEventTypeStart;
+
     Metroflip* app = context;
+    FuriString* parsed_data = furi_string_alloc();
+    SuicaHistoryViewModel* model = view_get_model(app->suica_context->view_history);
+
+    Widget* widget = app->widget;
+
     const FelicaPollerEvent* felica_event = event.event_data;
     FURI_LOG_I(TAG, "Felica event: %d", felica_event->type);
-
-    felica_event->data->auth_context->skip_auth = true;
-
-    if(felica_event->type == FelicaPollerEventTypeReady) {
-        SuicaHistoryViewModel* model = view_get_model(app->suica_context->view_history);
-        Widget* widget = app->widget;
-        FuriString* parsed_data = furi_string_alloc();
-
+    if(felica_event->type == FelicaPollerEventTypeReady ||
+       felica_event->type == FelicaPollerEventTypeIncomplete) {
         dolphin_deed(DolphinDeedNfcRead);
 
         FURI_LOG_I(TAG, "Read complete");
@@ -421,39 +420,17 @@ static NfcCommand suica_poller_callback(NfcGenericEvent event, void* context) {
 
         metroflip_app_blink_stop(app);
 
-        uint32_t system_count = simple_array_get_count(felica_data->systems);
-        bool suica_found = false;
-        bool octopus_found = false;
-        uint8_t suica_system_index = 0;
-        uint8_t octopus_system_index = 0;
-        for(uint8_t i = 0; i < system_count; i++) {
-            FelicaSystem* system = simple_array_get(felica_data->systems, i);
-            if(system->system_code == JAPAN_IC_SYSTEM_CODE) {
-                suica_found = true;
-                suica_system_index = i;
-            } else if(system->system_code == OCTOPUS_SYSTEM_CODE) {
-                octopus_found = true;
-                octopus_system_index = i;
-            }
-        }
+        bool suica_found = suica_model_pack_data(model, felica_data, parsed_data);
 
         do {
-            if(suica_found) {
-                FelicaSystem* suica_system =
-                    simple_array_get(felica_data->systems, suica_system_index);
-                furi_string_printf(parsed_data, "\e#Japan Transit IC\n");
-                suica_model_pack_data(model, suica_system, parsed_data);
-                break;
-            } else if(octopus_found) {
-                FelicaSystem* octopus_system =
-                    simple_array_get(felica_data->systems, octopus_system_index);
-                suica_help_with_octopus(octopus_system, parsed_data);
-                break;
-            } else {
-                furi_string_printf(
-                    parsed_data,
-                    "\e#FeliCa\nSorry, unrecorded service code.\nPlease let the developers know and we will add support.");
-            }
+            if(suica_found) break;
+
+            bool octopus_found = suica_help_with_octopus(felica_data, parsed_data);
+            if(octopus_found) break;
+
+            furi_string_printf(
+                parsed_data,
+                "\e#FeliCa\nSorry, unrecorded service code.\nPlease let the developers know and we will add support.");
         } while(false);
 
         widget_add_text_scroll_element(widget, 0, 0, 128, 64, furi_string_get_cstr(parsed_data));
@@ -469,8 +446,9 @@ static NfcCommand suica_poller_callback(NfcGenericEvent event, void* context) {
             widget, GuiButtonTypeLeft, "Save", metroflip_save_widget_callback, app);
 
         view_dispatcher_switch_to_view(app->view_dispatcher, MetroflipViewWidget);
-        furi_string_free(parsed_data);
     }
+    furi_string_free(parsed_data);
+    command = NfcCommandStop;
     return command;
 }
 
@@ -567,7 +545,6 @@ static void suica_on_enter(Metroflip* app) {
         popup_set_icon(app->popup, 0, 3, &I_RFIDDolphinReceive_97x61);
         view_dispatcher_switch_to_view(app->view_dispatcher, MetroflipViewPopup);
 
-        nfc_scanner_alloc(app->nfc);
         app->poller = nfc_poller_alloc(app->nfc, NfcProtocolFelica);
         nfc_poller_start(app->poller, suica_poller_callback, app);
         FURI_LOG_I(TAG, "Poller started");
@@ -589,34 +566,11 @@ static void suica_on_enter(Metroflip* app) {
             bool suica_found = false;
             do {
                 if(is_stock_nfc_file) { // loading from stock NFC file (API 87.0+)
-                    FURI_LOG_I(TAG, "Loading from stock NFC file");
-                    uint32_t system_count = simple_array_get_count(felica_data->systems);
-                    bool octopus_found = false;
-                    uint8_t suica_system_index = 0;
-                    uint8_t octopus_system_index = 0;
-                    for(uint8_t i = 0; i < system_count; i++) {
-                        FelicaSystem* system = simple_array_get(felica_data->systems, i);
-                        if(system->system_code == JAPAN_IC_SYSTEM_CODE) {
-                            suica_found = true;
-                            suica_system_index = i;
-                        } else if(system->system_code == OCTOPUS_SYSTEM_CODE) {
-                            octopus_found = true;
-                            octopus_system_index = i;
-                        }
-                    }
+                    suica_found = suica_model_pack_data(model, felica_data, parsed_data);
+                    if(suica_found) break;
 
-                    if(suica_found) {
-                        FelicaSystem* suica_system =
-                            simple_array_get(felica_data->systems, suica_system_index);
-                        furi_string_printf(parsed_data, "\e#Japan Transit IC\n\n");
-                        suica_model_pack_data(model, suica_system, parsed_data);
-                        break;
-                    } else if(octopus_found) {
-                        FelicaSystem* octopus_system =
-                            simple_array_get(felica_data->systems, octopus_system_index);
-                        suica_help_with_octopus(octopus_system, parsed_data);
-                        break;
-                    }
+                    bool octopus_found = suica_help_with_octopus(felica_data, parsed_data);
+                    if(octopus_found) break;
                 } else { // loading from legacy saved file (pre API 87.0)
                     suica_found = true;
                     furi_string_printf(parsed_data, "\e#Japan Transit IC\n\n");